generator/requirements.txt
kafka-python-ng
Faker 